Changed name of movie list component and structure for it, moved components into new folder.
