+++ code/src/components/movieDetails.css
@@ -1,8 +1,3 @@
-.movie-overview {
-  height: 100%;
-  width: 100%;
-}
-
 .movie-details {
   background-size: cover;
   color: #fff;
